Some minor clean-ups
